feat: extract theme logic
