+++ assets/js/app.js
@@ -1,0 +1,9 @@
+import { initMenu } from './menu.js';
+import { initTheme } from './theme.js';
+import { initSidebar } from './sidebar.js';
+
+document.addEventListener('DOMContentLoaded', () => {
+    initTheme();
+    initMenu();
+    initSidebar();
+});
