+++ includes/head_common.php
@@ -32,7 +32,7 @@ require_once __DIR__ . '/env_loader.php';
 <link rel="stylesheet" href="/assets/css/torch_cursor.css">
 <link rel="preload" href="/assets/js/torch_cursor.js" as="script">
 <script defer src="/assets/js/torch_cursor.js"></script>
-<link rel="preload" href="/assets/js/main.js" as="script">
+<link rel="preload" href="/assets/js/app.js" as="script">
 <link rel="stylesheet" href="/assets/css/glow_filter.css">
 <link rel="stylesheet" href="/assets/css/custom-pointer.css">
 <link rel="stylesheet" href="/assets/css/mobile_contrast.css" media="(max-width: 768px)">
